#1172 fix build issue

+++ FlowCrypt/Functionality/Services/Local Contacts Service/LocalContactsProvider.swift
@@ -123,7 +123,10 @@ extension LocalContactsProvider: LocalContactsProviderType {
         let keys = try find(with: email)?
             .pubKeys
             .filter { $0.primaryFingerprint == fingerprint }
-        for key in keys ?? [] {
+        guard let keys = keys else {
+            return
+        }
+        for key in keys {
             try storage.write {
                 storage.delete(key)
             }
